Added new funtion to param retriever and corrected minor bug for PI compilation

+++ src/module/mod_sst.cpp
@@ -43,7 +43,7 @@ void process_pot_msg(const odas_ros::pot::ConstPtr& in_ros_msg, msg_pots_obj* in
     out_ros_msg.nTracks = out_msg->tracks->nTracks;
 
     int nTracks = out_ros_msg.nTracks;
-    out_ros_msg.ids = std::vector<long unsigned int>(nTracks);
+    out_ros_msg.ids = std::vector<long long unsigned int>(nTracks);
     out_ros_msg.data = std::vector<float>(nTracks * 3);
 
     for(int iTracks = 0; iTracks < nTracks; iTracks++) {
